+++ packages/twenty-front/src/modules/ui/layout/selectable-list/components/SelectableList.tsx
@@ -3,7 +3,6 @@ import { ReactNode, useEffect } from 'react';
 import { useSelectableListHotKeys } from '@/ui/layout/selectable-list/hooks/internal/useSelectableListHotKeys';
 import { useSelectableList } from '@/ui/layout/selectable-list/hooks/useSelectableList';
 import { SelectableListScope } from '@/ui/layout/selectable-list/scopes/SelectableListScope';
-import { useRecoilValue } from 'recoil';
 import { arrayToChunks } from '~/utils/array/arrayToChunks';
 import { isDefined } from '~/utils/isDefined';
 
@@ -15,7 +14,6 @@ type SelectableListProps = {
   onSelect?: (selected: string) => void;
   hotkeyScope: string;
   onEnter?: (itemId: string) => void;
-  selectFirstItem?: boolean;
 };
 
 export const SelectableList = ({
@@ -25,18 +23,11 @@ export const SelectableList = ({
   selectableItemIdArray,
   selectableItemIdMatrix,
   onEnter,
-  selectFirstItem = false,
 }: SelectableListProps) => {
   useSelectableListHotKeys(selectableListId, hotkeyScope);
 
-  const {
-    setSelectableItemIds,
-    setSelectableListOnEnter,
-    setSelectedItemId,
-    selectedItemIdState,
-  } = useSelectableList(selectableListId);
-
-  const selectedItemId = useRecoilValue(selectedItemIdState);
+  const { setSelectableItemIds, setSelectableListOnEnter, setSelectedItemId } =
+    useSelectableList(selectableListId);
 
   useEffect(() => {
     setSelectableListOnEnter(() => onEnter);
@@ -55,23 +46,12 @@ export const SelectableList = ({
 
     if (isDefined(selectableItemIdArray)) {
       setSelectableItemIds(arrayToChunks(selectableItemIdArray, 1));
-
-      if (
-        selectFirstItem &&
-        selectableItemIdArray.length > 0 &&
-        (!isDefined(selectedItemId) ||
-          !selectableItemIdArray.includes(selectedItemId))
-      ) {
-        setSelectedItemId(selectableItemIdArray[0]);
-      }
     }
   }, [
     selectableItemIdArray,
     selectableItemIdMatrix,
     setSelectableItemIds,
     setSelectedItemId,
-    selectFirstItem,
-    selectedItemId,
   ]);
 
   return (
